# [1529] Include only course counts on provider index