--- a/app/serializers/api/v2/serializable_provider.rb
+++ b/app/serializers/api/v2/serializable_provider.rb
@@ -57,7 +57,7 @@ module API
 
       has_many :courses do
         meta do
-          { count: @object.courses.size }
+          { count: @object.courses_count }
         end
       end
     end