### Context

Improving our performance by optimising out N+1 and slow db queries.

### Changes proposed in this pull request

Uses a bit of raw SQL on the provider model to only include the course counts.

This vastly improves response times, two-fold:

- Cuts the database/ActiveRecord time 2.5x, by not fetching the entire course information as part of the call, only the count
- Cuts the view/json rendering time 5x, because `JSONAPI::Serializable::Resource` seems to do unnecessary work when the entire course information is fetched in

Before:

```
Completed 200 OK in 519ms (Views: 481.4ms | ActiveRecord: 36.4ms)
Completed 200 OK in 649ms (Views: 608.0ms | ActiveRecord: 38.9ms)
Completed 200 OK in 513ms (Views: 473.4ms | ActiveRecord: 38.4ms)
Completed 200 OK in 636ms (Views: 594.4ms | ActiveRecord: 39.6ms)
Completed 200 OK in 535ms (Views: 492.8ms | ActiveRecord: 40.0ms)
```

After:

```
Completed 200 OK in 96ms (Views: 76.9ms | ActiveRecord: 17.0ms)
Completed 200 OK in 97ms (Views: 79.2ms | ActiveRecord: 16.6ms)
Completed 200 OK in 136ms (Views: 118.1ms | ActiveRecord: 16.3ms)
Completed 200 OK in 97ms (Views: 78.8ms | ActiveRecord: 16.3ms)
Completed 200 OK in 98ms (Views: 79.3ms | ActiveRecord: 16.8ms)
```

### Guidance to review

More tests forthcoming.

### Checklist

- [x] Make sure all information from the Trello card is in here
- [x] Attach to Trello card
- [x] Rebased master
- [x] Cleaned commit history
- [x] Tested by running locally