@tvararu 2019-05-24
@benilovj tests added, comment added :shipit: