@tvararu 2019-05-24
![](https://github.trello.services/images/mini-trello-icon.png) [Optimise organisations page on Publish](https://trello.com/c/xy5Qxbw5/1532-optimise-organisations-page-on-publish)